fix: changes to fix codacy issue
