+++ india_compliance/gst_india/api_classes/base.py
@@ -271,3 +271,27 @@ class BaseAPI:
 
 def get_public_ip():
     return requests.get("https://api.ipify.org").text
+
+
+def check_scheduler_status():
+    """
+    Throw an error if scheduler is disabled
+    """
+
+    if (
+        not frappe.conf.developer_mode
+        and frappe.utils.scheduler.is_scheduler_disabled()
+    ):
+        frappe.throw(
+            _(
+                "The Scheduler is currently disabled, which needs to be enabled to use e-Invoicing and e-Waybill features. "
+                "Please get in touch with your server administrator to resolve this issue.<br><br>"
+                "For more information, refer to the following documentation: {0}"
+            ).format(
+                """
+                <a href="https://frappeframework.com/docs/user/en/bench/resources/bench-commands-cheatsheet#scheduler" target="_blank">
+                    Frappe Scheduler Documentation
+                </a>
+                """
+            )
+        )
